refactor(divide): apply divide to conquer for smaller component
